Exercices/TP1/requirements.txt
mysql-connector-python==9.3.0
dotenv==0.9.9
tabulate==0.9.0